+++ tests/integration/conftest.py
@@ -1,3 +1,7 @@
+import os
+import pathlib
+
+import earthaccess
 import pytest
 
 ACCEPTABLE_FAILURE_RATE = 10
@@ -29,3 +33,37 @@ def pytest_sessionfinish(session, exitstatus):
     failure_rate = (100.0 * session.testsfailed) / session.testscollected
     if failure_rate <= ACCEPTABLE_FAILURE_RATE:
         session.exitstatus = 99
+
+
+@pytest.fixture
+def mock_env(monkeypatch):
+    earthaccess.__auth__ = earthaccess.Auth()
+    # the original comes from github secrets
+    monkeypatch.setenv("EARTHDATA_USERNAME", os.getenv("EARTHACCESS_TEST_USERNAME", ""))
+    monkeypatch.setenv("EARTHDATA_PASSWORD", os.getenv("EARTHACCESS_TEST_PASSWORD", ""))
+
+
+@pytest.fixture
+def mock_missing_netrc(tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch):
+    netrc_path = tmp_path / ".netrc"
+    monkeypatch.setenv("NETRC", str(netrc_path))
+    monkeypatch.delenv("EARTHDATA_USERNAME")
+    monkeypatch.delenv("EARTHDATA_PASSWORD")
+    # Currently, due to there being only a single, global, module-level auth
+    # value, tests using different auth strategies interfere with each other,
+    # so here we are deleting the auth attribute so that it doesn't interfere.
+    monkeypatch.delattr(earthaccess, "__auth__", raising=False)
+
+
+@pytest.fixture
+def mock_netrc(mock_env, tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch):
+    netrc = tmp_path / ".netrc"
+    monkeypatch.setenv("NETRC", str(netrc))
+
+    username = os.environ["EARTHDATA_USERNAME"]
+    password = os.environ["EARTHDATA_PASSWORD"]
+
+    netrc.write_text(
+        f"machine urs.earthdata.nasa.gov login {username} password {password}\n"
+    )
+    netrc.chmod(0o600)