+++ tests/integration/test_cloud_open.py
@@ -1,8 +1,5 @@
-# package imports
 import logging
-import os
 import random
-import unittest
 
 import earthaccess
 import magic
@@ -55,20 +52,6 @@ daacs_list = [
     },
 ]
 
-assertions = unittest.TestCase("__init__")
-
-# we need to use a valid EDL credential
-
-assertions.assertTrue("EARTHDATA_USERNAME" in os.environ)
-assertions.assertTrue("EARTHDATA_PASSWORD" in os.environ)
-
-auth = Auth().login(strategy="environment")
-assertions.assertTrue(auth.authenticated)
-logger.info(f"Current username: {os.environ['EARTHDATA_USERNAME']}")
-logger.info(f"earthaccess version: {earthaccess.__version__}")
-
-store = Store(auth)
-
 
 def get_sample_granules(granules, sample_size, max_granule_size):
     """Returns a list with sample granules and their size in MB if
@@ -82,11 +65,9 @@ def get_sample_granules(granules, sample_size, max_granule_size):
     while tries <= max_tries:
         g = random.sample(granules, 1)[0]
         if g.size() > max_granule_size:
-            # print(f"G: {g['meta']['concept-id']} exceded max size: {g.size()}")
             tries += 1
             continue
         else:
-            # print(f"Adding : {g['meta']['concept-id']} size: {g.size()}")
             files_to_download.append(g)
             total_size += g.size()
             if len(files_to_download) >= sample_size:
@@ -95,14 +76,11 @@ def get_sample_granules(granules, sample_size, max_granule_size):
 
 
 def supported_collection(data_links):
-    for url in data_links:
-        if "podaac-tools.jpl.nasa.gov/drive" in url:
-            return False
-    return True
+    return all("podaac-tools.jpl.nasa.gov/drive" not in url for url in data_links)
 
 
 @pytest.mark.parametrize("daac", daacs_list)
-def test_earthaccess_can_open_onprem_collection_granules(daac):
+def test_earthaccess_can_open_onprem_collection_granules(mock_env, daac):
     """Tests that we can download cloud collections using HTTPS links."""
     daac_shortname = daac["short_name"]
     collections_count = daac["collections_count"]
@@ -115,17 +93,18 @@ def test_earthaccess_can_open_onprem_collection_granules(daac):
     hits = collection_query.hits()
     logger.info(f"Cloud hosted collections for {daac_shortname}: {hits}")
     collections = collection_query.get(collections_count)
-    assertions.assertGreater(len(collections), collections_sample_size)
+    assert len(collections) > collections_sample_size
     # We sample n cloud hosted collections from the results
     random_collections = random.sample(collections, collections_sample_size)
     logger.info(f"Sampled {len(random_collections)} collections")
+
     for collection in random_collections:
         concept_id = collection.concept_id()
         granule_query = DataGranules().concept_id(concept_id)
         total_granules = granule_query.hits()
         granules = granule_query.get(granules_count)
-        assertions.assertTrue(len(granules) > 0, "Could not fetch granules")
-        assertions.assertTrue(isinstance(granules[0], earthaccess.results.DataGranule))
+        assert len(granules) > 0, "Could not fetch granules"
+        assert isinstance(granules[0], earthaccess.DataGranule)
         data_links = granules[0].data_links()
         if not supported_collection(data_links):
             logger.warning(f"PODAAC DRIVE is not supported at the moment: {data_links}")
@@ -143,10 +122,11 @@ def test_earthaccess_can_open_onprem_collection_granules(daac):
             f"download size(MB): {total_size_cmr}"
         )
 
+        store = Store(Auth().login(strategy="environment"))
         # We are testing this method
         fileset = store.open(granules_to_open)
 
-        assertions.assertTrue(isinstance(fileset, list))
+        assert isinstance(fileset, list)
 
         # we test that we can read some bytes and get the file type
         for file in fileset:
@@ -163,4 +143,4 @@ def test_multi_file_granule():
     urls = granules[0].data_links()
     assert len(urls) > 1
     files = earthaccess.open(granules)
-    assert set(urls) == set(f.path for f in files)
+    assert set(urls) == {f.path for f in files}